Rename `prevent-abbreviations` to `name-replacements`
