+++ readme.md
@@ -96,6 +96,7 @@ export default defineConfig([
 | [isolated-functions](docs/rules/isolated-functions.md)                                                     | Prevent usage of variables from outside the scope of isolated functions.                                                       | ✅    |    |    |    |
 | [logical-assignment-operators](docs/rules/logical-assignment-operators.md)                                 | Require or disallow logical assignment operator shorthand                                                                      | ✅    | 🔧 | 💡 |    |
 | [max-nested-calls](docs/rules/max-nested-calls.md)                                                         | Limit the depth of nested calls.                                                                                               | ✅    |    |    |    |
+| [name-replacements](docs/rules/name-replacements.md)                                                       | Enforce replacements for variable, property, and filenames.                                                                    | ✅    | 🔧 | 💡 |    |
 | [new-for-builtins](docs/rules/new-for-builtins.md)                                                         | Enforce correct use of `new` for builtin constructors.                                                                         | ✅ ☑️ | 🔧 | 💡 |    |
 | [no-abusive-eslint-disable](docs/rules/no-abusive-eslint-disable.md)                                       | Enforce specifying rules to disable in `eslint-disable` comments.                                                              | ✅ ☑️ |    |    |    |
 | [no-accessor-recursion](docs/rules/no-accessor-recursion.md)                                               | Disallow recursive access to `this` within getters and setters.                                                                | ✅ ☑️ |    |    |    |
@@ -341,7 +342,6 @@ export default defineConfig([
 | [prefer-unicode-code-point-escapes](docs/rules/prefer-unicode-code-point-escapes.md)                       | Prefer Unicode code point escapes over legacy escape sequences.                                                                | ✅ ☑️ | 🔧 | 💡 |    |
 | [prefer-url-href](docs/rules/prefer-url-href.md)                                                           | Prefer `URL#href` over stringifying a `URL`.                                                                                   | ✅ ☑️ | 🔧 |    |    |
 | [prefer-while-loop-condition](docs/rules/prefer-while-loop-condition.md)                                   | Prefer putting the condition in the while statement.                                                                           | ✅ ☑️ | 🔧 |    |    |
-| [prevent-abbreviations](docs/rules/prevent-abbreviations.md)                                               | Prevent abbreviations.                                                                                                         | ✅    | 🔧 | 💡 |    |
 | [relative-url-style](docs/rules/relative-url-style.md)                                                     | Enforce consistent relative URL style.                                                                                         | ✅ ☑️ | 🔧 | 💡 |    |
 | [require-array-join-separator](docs/rules/require-array-join-separator.md)                                 | Enforce using the separator argument with `Array#join()`.                                                                      | ✅ ☑️ | 🔧 |    |    |
 | [require-array-sort-compare](docs/rules/require-array-sort-compare.md)                                     | Require a compare function when calling `Array#sort()` or `Array#toSorted()`.                                                  | ✅ ☑️ |    | 💡 |    |
