+++ rules/index.js
@@ -33,6 +33,7 @@ export {default as 'import-style'} from './import-style.js';
 export {default as 'isolated-functions'} from './isolated-functions.js';
 export {default as 'logical-assignment-operators'} from './logical-assignment-operators.js';
 export {default as 'max-nested-calls'} from './max-nested-calls.js';
+export {default as 'name-replacements'} from './name-replacements.js';
 export {default as 'new-for-builtins'} from './new-for-builtins.js';
 export {default as 'no-abusive-eslint-disable'} from './no-abusive-eslint-disable.js';
 export {default as 'no-accessor-recursion'} from './no-accessor-recursion.js';
@@ -278,7 +279,6 @@ export {default as 'prefer-unary-minus'} from './prefer-unary-minus.js';
 export {default as 'prefer-unicode-code-point-escapes'} from './prefer-unicode-code-point-escapes.js';
 export {default as 'prefer-url-href'} from './prefer-url-href.js';
 export {default as 'prefer-while-loop-condition'} from './prefer-while-loop-condition.js';
-export {default as 'prevent-abbreviations'} from './prevent-abbreviations.js';
 export {default as 'relative-url-style'} from './relative-url-style.js';
 export {default as 'require-array-join-separator'} from './require-array-join-separator.js';
 export {default as 'require-array-sort-compare'} from './require-array-sort-compare.js';
