+++ test/integration/fixtures-local/conflicts-no-array-for-each-and-name-replacements.js
