+++ test/name-replacements.js
